IndexedDB/tests/submissions/Opera/keygenerator2.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>Keygenerator ConstraintError</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>

    var db,
      t = async_test(),
      objects = [1, null, {id: 2}, null, 2.00001, 5, null ],
      expected = [1, 2, 2.00001, 3, 5, 6];

    var open_rq = createdb(t);
    open_rq.onupgradeneeded = function(e) {
        db = e.target.result;
        var objStore = db.createObjectStore("store", { keyPath: "id", autoIncrement: true });

        for (var i = 0; i < objects.length; i++)
        {
            if (objects[i] === null)
            {
                objStore.add({});
            }
            else if (typeof objects[i] === "object")
            {
                assert_throws("ConstraintError",
                    function() { objStore.add(objects[i]) });
            }
            else
                objStore.add({ id: objects[i] });
        }
    };

    open_rq.onsuccess = function(e) {
        var actual_keys = [],
          rq = db.transaction("store")
                 .objectStore("store")
                 .openCursor();

        rq.onsuccess = t.step_func(function(e) {
            var cursor = e.target.result;

            if (cursor) {
                actual_keys.push(cursor.key.valueOf());
                cursor.continue();
            }
            else {
                assert_equals(actual_keys.length, expected_keys.length, "array length");
                assert_object_equals(actual_keys, expected_keys, "keygenerator array");

                t.done();
            }
        });
    };

</script>

<div id="log"></div>
